refactor: split browser detection into src/browser.ts
index.ts shrinks 765 -> 356 lines. Browser detection, URL building,
and spawn helpers move to a focused module. Public API is preserved
via a re-export from index.ts so existing consumers and tests keep
working unchanged.
